Send stdin through a pipe when attaching
When the connection to the Docker server drops (e.g. due to the
container exiting), fsouza/go-dockerclient closes the input stream we
provided [1]. io.Copy is used to copy the input stream to the docker
server [2], and AttachToContainer only returns once that io.Copy stops
blocking. However, closing os.Stdin doesn't have an immediate effect on
io.Copy (but you do get an error next time stdin receives any data).

To get around this issue, we pass fsouza/go-dockerclient the reader side
of an io.Pipe, and in a separate goroutine copy from stdin to the writer
half of the pipe. When fsouza/go-dockerclient closes the reader half of
the pipe, the io.Copy call at [2] immediately unblocks, letting
AttachToContainer return.

A caveat is that we'll leak the goroutine that copies from stdin to the
pipe until stdin receives more input (when it'll cause a read error that
is left unhandled).

The fundamental problem here is that we have no way of interrupting the
call to stdin.Read() (from within io.Copy) from another goroutine when
the connection is severed.

[1]: https://github.com/fsouza/go-dockerclient/blob/1399676f53e6ccf46e0bf00751b21bed329bc60e/client.go#L621
[2]: https://github.com/fsouza/go-dockerclient/blob/1399676f53e6ccf46e0bf00751b21bed329bc60e/client.go#L636